requirements.txt
fastapi
uvicorn
pandas
scikit-learn
prophet
yfinance
python-jose
passlib[bcrypt]
sqlalchemy